New Feature: EVEMon now has a character Asset monitor. Additionally laid code base for upcoming features. Many improvements have been made to the listview updating logic of a monitor (to each monitor) to increase the speed of populating.

+++ EVEMon.Common/Data/Material.cs
@@ -36,7 +36,7 @@ namespace EVEMon.Common.Data
         /// <summary>
         /// Gets the reprocessed quantity.
         /// </summary>
-        public int Quantity { get; private set; }
+        public long Quantity { get; private set; }
 
         #endregion
     }
